fixed unittest problems
